SAP: Detect when SAP Announcements time out
This is not how RFC 2974 says it should be done, and TBH, the method
described as in RFC 2974 probably is not a good match for current day
networks and the actual requirements. The device we're testing with
announces every 30.5 seconds, so we will assume announcements are due
every 30 seconds, and will consider announcements timed out when they
have been missed 10 times, i.e. after 5 minutes.